Create IQ1.java
This can be used as an API for doing Immutable Operations of Concurrent Updates..
In this programm UNNECESSARY COPY has been removed..
But Still COPYING is to be done while Enqueuing in Worst Case..
There is also FULL_IMMUTABLE.java file in which COPYING has been replaced by a sequence of Dequeue and Enqueue i.e. maintaining our queue as Completely Immutable.